+++ CoreCoderSharp/UI/TuiCust/UxHelper.cs
@@ -1,6 +1,6 @@
 using CoreCoderSharp.Terminal;
 using CoreCoderSharp.UI.Controls;
-using CoreCoderSharp.UI.TuiBase;
+
 using CoreCoderSharp.UI.TuiControls;
 
 namespace CoreCoderSharp.UI;